Remove Backend trait and turn Module into a trait
